+++ packages/expo/src/dom/marshal.tsx
@@ -1,16 +1,14 @@
 import { BridgeMessage, JSONValue } from './dom.types';
 import { DOM_EVENT, NATIVE_ACTION, NATIVE_ACTION_RESULT } from './injection';
 
-const IS_DOM =
-  typeof window !== 'undefined' &&
-  // @ts-expect-error: Added via react-native-webview
-  typeof window.ReactNativeWebView !== 'undefined';
+const IS_DOM = typeof window !== 'undefined' && window.isDOMComponentContext === true;
 
 const emit = <TData extends JSONValue>(message: BridgeMessage<TData>) => {
   if (!IS_DOM) {
     return;
   }
-  (window as any).ReactNativeWebView.postMessage(JSON.stringify(message));
+  // @ts-expect-error: Added via react-native-webview
+  window.ReactNativeWebView.postMessage(JSON.stringify(message));
 };
 
 export const addEventListener = <TData extends JSONValue>(